Added Blog Functions

+++ routes/api.php
@@ -7,17 +7,17 @@ use App\Http\Controllers\PostController;
 // 🔐 AUTH
 Route::post('/login', [AuthController::class, 'login']);
 
-// 🌍 PUBLIC BLOG
+// 🌍 PUBLIC BLOG (Csak a látogatóknak)
 Route::get('/posts', [PostController::class, 'index']);
 Route::get('/posts/{post}', [PostController::class, 'show']);
 
-// 🔐 ADMIN CMS
+// 🔐 ADMIN CMS (Bejelentkezett felhasználóknak)
 Route::middleware('auth:sanctum')->group(function () {
-
-    Route::post('/posts', [PostController::class, 'store']);
-    Route::put('/posts/{post}', [PostController::class, 'update']);
-    Route::delete('/posts/{post}', [PostController::class, 'destroy']);
+    Route::get('/admin/posts', [PostController::class, 'adminIndex']);
+    Route::post('/admin/posts', [PostController::class, 'store']);
+    Route::put('/admin/posts/{post}', [PostController::class, 'update']);
+    Route::delete('/admin/posts/{post}', [PostController::class, 'destroy']);
 
     // 🖼️ IMAGE UPLOAD
-    Route::post('/upload', [PostController::class, 'upload']);
+    Route::post('/admin/upload', [PostController::class, 'upload']);
 });
